refactor the settings screen to use a nested NavHostGraph
