+++ frontend/android/app/src/main/java/cs/vsu/taskbench/ui/settings/SettingsGraph.kt
@@ -1,0 +1,6 @@
+package cs.vsu.taskbench.ui.settings
+
+import com.ramcosta.composedestinations.annotation.NavHostGraph
+
+@NavHostGraph
+annotation class SettingsGraph
