+++ frontend/android/app/src/main/java/cs/vsu/taskbench/ui/settings/SettingsMenuOption.kt
@@ -1,0 +1,48 @@
+package cs.vsu.taskbench.ui.settings
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import cs.vsu.taskbench.ui.theme.DarkGray
+
+@Composable
+fun SettingsMenuOption(
+    text: String,
+    icon: Painter,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = modifier
+            .clickable(onClick = onClick)
+            .height(52.dp)
+            .fillMaxWidth(),
+    ) {
+        Icon(
+            painter = icon,
+            contentDescription = null,
+            tint = DarkGray,
+            modifier = Modifier.size(32.dp),
+        )
+        Text(
+            text = text,
+            fontSize = 18.sp,
+            fontWeight = FontWeight.Medium,
+            color = DarkGray,
+        )
+    }
+}
